applications/testautogen/pyproject.toml
[project]
name = "testautogen"
version = "2024.0.0"
description = "testautogen"
readme = "README.md"
requires-python = ">=3.10"
license = { file = "LICENSE.txt" }
keywords = ["Artificial Intelligence", "Machine Learning"]
authors = [{ email = "tom.wieland@gmail.com" }, { name = "Tom Wieland" }]
maintainers = [{ name = "Tom Wieland", email = "tom.wieland@gmail.com" }]
classifiers = [
  "Development Status :: 4 - Beta",
  "Programming Language :: Python",
]
dependencies = [
  "black>23.3.0",
  "debugpy>1.6.7",
  "flake8>6.0.0",
  "isort>=5.12.0",
  # "pylance>0.9.8",

  "pytest",
  "pytest-cov",
  "pytest-html",

  "python-dotenv",

  "chromadb",
  # "flaml",
  "pypdf",
  "sentence_transformers",

  "anyio",
  "diskcache",
  "distro",
  "docker",
  "flaml",
  "hatchling",
  "httpx",
  "numpy",
  "openai",
  "pyautogen",
  "pyautogen[retrievechat]",
  "pydantic",
  "unstructured",
  "unstructured[all-docs]",
  "python-dotenv",
  "setuptools",
  "sniffio",
  "termcolor",
  "tiktoken",
  "tqdm",
  "typing-extensions",
  "ipython",
]

[tool.coverage.run]
branch = true
source = ["testautogen"]

[tool.coverage.report]
exclude_lines = ['if TYPE_CHECKING:']
show_missing = true

[tool.pytest.ini_options]
addopts = "--cov --cov-report html:'../../coverage/applications/testautogen/html' --cov-report xml:'../../coverage/applications/testautogen/coverage.xml' --html='../../reports/applications/testautogen/unittests/html/index.html' --junitxml='../../reports/applications/testautogen/unittests/junit.xml'"

[tool.nx]
autoActivate = true
